Rewrote VGUI to prevent unwanted charge bars from appearing when holding right click with a two-handed weapon or holding a single handed weapon.
Also fixes the right click charge also showing an erroneous main hand charge bar while dual wieldin.

This is a bit dumb and I don't really like it but it kept crashing if I tried using references so we are just going to have a vector of pointers that gets deleted a lot.