feat: refactor funcmonsterclip and add funcplayerclip (#212)
* feat: refactor funcmonsterclip

Changed CFuncMonsterClip to be a CFuncWallToggle (to mask it in case any engine level code is looking for FuncWallToggles) with relevant overrides. Works like a func_wall_toggle for monsters only without the hacky origin teleportation to turn on and off. Always passable to players.

* feat: add msfuncplayerclip

Added CMSFuncPlayerClip, which works like a func_wall_toggle. If the player touches it, it pushes them back scaled to the entity's dmg value as set by the mapper. Monsters can pass right through.